simplified operators related code
