+++ src/main/java/com/assignment/rpn/operator/Operators.java
@@ -29,40 +29,12 @@ public final class Operators {
     private Operators() {
     }
 
-    public static final Operator<BigDecimal> ADD = new AbstractOperator<>("+") {
-        @Override
-        public void apply(OperatorContext<BigDecimal> context) throws InsufficientParametersException {
-            apply(BigDecimal::add, context);
-        }
-    };
-
-    public static final Operator<BigDecimal> SUBTRACT = new AbstractOperator<>("-") {
-        @Override
-        public void apply(OperatorContext<BigDecimal> context) throws InsufficientParametersException {
-            apply(BigDecimal::subtract, context);
-        }
-    };
-
-    public static final Operator<BigDecimal> MULTIPLY = new AbstractOperator<>("*") {
-        @Override
-        public void apply(OperatorContext<BigDecimal> context) throws InsufficientParametersException {
-            apply(BigDecimal::multiply, context);
-        }
-    };
-
-    public static final Operator<BigDecimal> DIVIDE = new AbstractOperator<>("/") {
-        @Override
-        public void apply(OperatorContext<BigDecimal> context) throws InsufficientParametersException {
-            apply(BigDecimal::divide, context);
-        }
-    };
-
-    public static final Operator<BigDecimal> SQRT = new AbstractOperator<>("sqrt") {
-        @Override
-        public void apply(OperatorContext<BigDecimal> context) throws InsufficientParametersException {
-            apply(bigDecimal -> bigDecimal.sqrt(Operators.DECIMAL_CONTEXT), context);
-        }
-    };
+    public static final Operator<BigDecimal> ADD = new BinaryOperator<>("+", BigDecimal::add);
+    public static final Operator<BigDecimal> SUBTRACT = new BinaryOperator<>("-", BigDecimal::subtract);
+    public static final Operator<BigDecimal> MULTIPLY = new BinaryOperator<>("*", BigDecimal::multiply);
+    public static final Operator<BigDecimal> DIVIDE = new BinaryOperator<>("/", BigDecimal::divide);
+    public static final Operator<BigDecimal> SQRT = new UnaryOperator<>(
+            "sqrt", d -> d.sqrt(Operators.DECIMAL_CONTEXT));
 
     public static final Operator<BigDecimal> CLEAR = new AbstractOperator<>("clear") {
         @Override
@@ -85,4 +57,11 @@ public final class Operators {
             }
         }
     };
+
+    public static void checkSufficientOperands(int requiredNumberOfOperands, int stackSize)
+            throws InsufficientParametersException {
+        if (stackSize < requiredNumberOfOperands) {
+            throw new InsufficientParametersException("insufficient parameters");
+        }
+    }
 }
